Tutorial Update & Game UI
Still quite a bit of bugs with some of the movements, will come back later and clean that up. The tutorial section has added quite a bit, but will likely need to have the code refactored after as its getting pretty messy. Game UI updates for piece "tooltips" as well as ability button that was needed for Jade tutorial.