Fixed Portal Bug
Not Ideal - and likely should do a revamp so that portals are shown as an option without needing to click on the piece - as well as showing indicators of possible swap positions...but I think I need to rework the structure of my code. :/

For now, at least this seems to work.